requirements.txt
pipDjango==2.0
django-crispy-forms==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
PyJWT==1.7.1
pytz==2019.3

